+++ module/card/card-helpers.js
@@ -1,6 +1,6 @@
 import { HeroSystem6eCard } from "./card.js";
 import { HeroSystem6eAttackCard } from "./attack-card.js";
-import { HeroSystem6eToHitCard2 } from "./toHit-card2.js"
+import { HeroSystem6eToHitCard2 } from "./toHit-card.js"
 import * as Attack from "../item/item-attack.js"
 
 export class HeroSystem6eCardHelpers {