+++ module/template.js
@@ -1,3 +1,0 @@
-export default class HeroSystem6eTemplate extends MeasuredTemplate {
-
-}